+++ packages/desktop-client/src/components/spreadsheet/index.ts
@@ -1,14 +1,29 @@
-// @ts-strict-ignore
-import {
-  type SheetNames,
-  type BudgetField,
-  type SpreadsheetFieldTypes,
-} from 'loot-core/client/queries';
 import { type Query } from 'loot-core/src/shared/query';
+
+export type SpreadsheetFieldTypes = {
+  account: {
+    // Common fields
+    'uncategorized-amount': number;
+    'uncategorized-balance': number;
+
+    // Account fields
+    balance: number;
+    'accounts-balance': number;
+    'budgeted-accounts-balance': number;
+    'offbudget-accounts-balance': number;
+    balanceCleared: number;
+    balanceUncleared: number;
+  };
+};
+
+export type SheetNames = keyof SpreadsheetFieldTypes & string;
+
+export type SheetFields<SheetName extends SheetNames> =
+  keyof SpreadsheetFieldTypes[SheetName] & string;
 
 export type Binding<
   SheetName extends SheetNames = any,
-  FieldName extends BudgetField<SheetName> = any,
+  FieldName extends SheetFields<SheetName> = any,
 > =
   | FieldName
   | {
